Add exhibit3 library; only create exhibit after scripts have loaded
First, we had to find out that ahbeng's NUSMods repository was making use of
this library. Then, many hours were spent studying the exhibit3 library and
some of its demos. Thereafter, figuring out how to create the exhibit after
the scripts have loaded, which was accomplished with the hint in this
link:

    http://osdir.com/ml/web.simile.general/2007-11/msg00106.html

and studying the related section of this file:

    https://github.com/ahbeng/NUSMods/blob/master/src/coffee/main.coffee

exhibit 3.0 is a very powerful library which is seriously lacking in
documentation, as is the case with many libraries out there.
I must write a tutorial on it once this project reaches a more stable state
and when I'm more familiar with it.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>