[INVESTIGATION] Figure out keys and some types in modules.json
We are using module data from here:

    http://api.nusmods.com/

to construct module objects in our code. One such file is:

    http://api.nusmods.com/2013-2014/2/modules.json

However, some investigation is required to understand more about the individual
module objects in the `modules.json` file. In this patch, we introduce one
program which helps us to do so, the `inspect_scripts/types.js` file.
This script gives a high level overview of the different keys in each module
object, along with the types of their values.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>