[INVESTIGATION] Add script for value analysis (for keys with string values)
Here, value analysis refers to a simple summary of the number of times a certain
value appears, say X. We're interested in finding out for a given key, how many
unique values appear X times. We call X a "bucket".

For keys that have a small amount of buckets and a large unique amount of
strings for those buckets (say we have 1000 strings that appear 500 times each),
then we might not want to store each of those strings X times for space
efficiency.

While this script is still very raw, it will be polished in upcoming patches to
give us better insight for our data representation for the `modules.json` file.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>