[INVESTIGATION] String format of ModuleCode
This is an investigation into the format of module codes. It is done to prepare
for an upcoming patch for enabling filtering modules by Level.

It is suspected that all module codes follow this format:

    <1 or more non digit chars><4 digit chars><0 or more non digit chars>

It is too tedious to go through all 2894 module codes manually, so a computer
program is written to ease this tedium.

It turns out that our theory is correct, except for the following module code:

    PH2302 / GEK2039

which must be handled specially when constructing the module levels.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>