[Bugfix] Use uppercase search term for searching
We use a custom `query` function for the select2 library for searching and
inside this function, the haystack string is uppercased, but the needle (search
term) is still in its original case. This results in lowercase searches yielding
no results.

This patch fixes the problem by using an uppercase'd needle string.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>