main.cljs - wrap page initialization in `init` function
This is to prepare for an upcoming patch which tests some of our ClojureScript
code. Previously, code in `main.cljs` was executed unconditionally, which is
not desirable in tests. Wrapping the main program initialization code in
functions and including the single line of:

    cljs_nusmods.main.init();

in the `index.html` file will prevent the same code from being executed unless
we explicitly allow it to.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>