Abstract cljs-nusmods.module-array-repr/get-module-lecture-timings
We create a function maker in
`module-lesson-timings-for-exhibit-filter-fn-maker`, which takes in a
single String parameter for the lesson type. The returned function does
the same thing as `get-module-lecture-timings`, with the difference of
that the lesson type String is supplied instead of hardcoded as
'Lecture'.

This is to prepare for the upcoming patch where we will implement a
`get-module-tutorial-timings` function for retrieving a Module object's
tutorial timings.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>