Use 1 Select2 input element for `Search Modules for Timetable` box
This avoids the complications required to synchronize the selected
values in 2 boxes. To use 1 Select2 input element across the 2 tabs, we
give an id to its container div, and move that div to the currently
selected tab using jQuery.prototype.prepend. The currently selected tab
is detected using a click event handler.

Some additional changes include:
- Adding several helper functions in the cljs-nusmods.select2 namespace
- For the `initSelection` function passed to the Select2 constructor,
  empty strings are filtered out of anything passed through the `.val`
  method

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>