Add background colors for lesson divs
We choose 24 colors for the timetable lesson divs, each with a css class
ending with the digits '00' to '23'. When time comes to allocate a
background color for a selected module, shuffled sequence of the integers
0 to 23 is generated. We pick the background color css class from this
sequence until it runs out of elements (when that happens, a new
shuffled sequence of 0 to 23 is generated, etc).

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>