main.cljs - Add 'test' code
This is code we use to check if things are working. It'd seem very
dishonest not to add this in and leave anyone who happens to be browsing
the history of this repository confused.

So here's a slightly more honest picture of how things are like at the
moment. There's more tidying up I need to do, especially with regards to
building the repository.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>