Add inspect_scripts/classno_format.js
It's been a while since we've done any inspection of module information.
This is a new inspect script that checks if all lesson group labels (the
`.ClassNo` attribute in Lesson objects) match a given regex in the code.

The purpose of this script is to help us determine a suitable regular
expression for matching all lesson group labels.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>