Add module lesson groups from url hash during initialization
This is similar to the feature available in NUSMods, except that it is
more clear, due to the distinction between different types of lessons
(such as lectures, tutorials, labs, etc) for each module lesson group
present in the url hash.

The addition of this patch allows us to make the
`add-module-lesson-group` private to its namespace, and replace the
lengthy and hackish addition of modules in main.cljs with this url hash:

    #ACC1002_L=V1&ACC1002_T=V01&CS1010_ST=1&CS1010_T=1&CS1020_LAB=1&CS1020_ST=1&CS1020_T=2&CM1121_L=SL1&CM1121_T=TH1&BPS5103_L=1&BPS5103_T=1&AR4223_L=1&AR4223_T=1&DSC3226_ST=Z1

However, this patch is not entirely complete. Here are several other
additions, some of which we will make, and some of which we may not:

- Handle missing lesson groups for a module
- Handle duplicate lesson groups for a module
- Report any lesson group specified in an invalid format, any missing
  module / lesson group

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>