Deal with missing lesson types for modules in url hash
The user may have missed out on some lesson types (eg. Tutorial, Lab)
for certain modules in the url hash. We randomly pick a lesson group
for that missing lesson type to present a full picture of the module.
The user can later switch the lesson group for that lesson type as
he/she deems fit.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>