Change data representation of `Timetable` and `ModulesSelected` globals
Some progress was made in implementing the removal of modules on a
separate branch, and along the way, we discovered some critical
weaknesses of our current data representation, in particular for the
`Timetable` global.

This patch introduces some changes to the `Timetable` and
`ModulesSelected` globals. For the `ModulesSelected` global, we remove
the `:divElem` key, instead forcing our functions to lookup the
appropriate day and row in the `Timetable` global for the actual jQuery
<div> object representing a lesson.

For the `Timetable` global, we abolish the use of a vector for marking
and searching for timeslots. Instead, we use a map to represent a row in
a day. Each key of this map is a `TimetableLessonInfo` data type defined
at the start of the `src/cljs/cljs_nusmods/timetable.cljs` file, and
each value is the jQuery <div> object for that lesson.

This will hopefully lead to a cleaner implementation when we implement
the remaining code required for removing modules.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>