Call `select2-box-update-modules` once for modules added via url hash initialization
Previously, the `select2-box-update-modules` function was called for
each call to `add-module-lesson-group!`. This resulted in redundant work
being done for updating the Select2 box. To illustrate what is meant by
redundant work.

When module A is added, the following modules gets processed by the
Select2 functions:

    A

When a new module B is added, the following modules get processed by the
Select2 functions:

    A B

When a new module C is added, the following modules get processed by the
Select2 functions:

    A B C

This is prevented in the current patch, where we add an
`addedViaUrlHashInit?` keyword parameter to the
`add-module-lesson-group!` function to control whether an update to the
Select2 box should be performed.

By doing this update only once during initialization of the application,
we save up on a good amount of work.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>