Lesson group removal - remove from `Timetable` first
Previously, We relied on the return value of the
`remove-lesson-group-html` function to find out the lessons we need to
remove from the `Timetable`. This defeats our objective of using the
`Timetable` as the single source of truth for the current state of
Module selection.

This patch changes things so that for lesson group removal, changes are
first made in `Timetable` and the lessons removed are then used as a
basis to make the corresponding changes in the HTML Timetable.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>