[STYLE] Select2 input box - add box shadows
We add the box shadows to the <ul> with class `select2-choices`, since
it is the element with the original Select2 border styles.

In addition, the actual <input> element shrinks whenever the Select2
menu moves from the menu open to menu close states, and adding border
styles to the <input> element causes this shrinking to become extremely
clear in a very ugly way.

On the ClojureScript side, we needed to add the
`select2-container-active` class to the Select2 box for the box shadows
to remain present when the select-close event is fired.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>