[STYLE] Remove margin-top for .select-results
For the `Timetable` page, whenever we use the Chromium browser to carry
out the following actions:

- Click on the Select2 box and click away
- Drag a draggable module down towards the Select2 box

A very obvious light blue border will jut out. Part of what makes this
light blue border an eyesore is the margin-top on the <ul> with the
`select-results` CSS class.

In this patch, we remove the margin-top for the `select-results` CSS
class. While this does not solve the problem, it does make the border
less of an eyesore.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>