[REFACTOR] Do not supply url hash for adding modules via in main.cljs
This is done in the reworked `add-module-lesson-groups-from-url-hash!`
function, which is a small wrapper over the original function of the
same name (now prefixed with `real-`). This is to prepare for the
upcoming patches, where we will use the url hash from localStorage as
a fallback from an empty `document.location.hash`.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>