Use /acad-year/sem/ for route prefix
This is the start of a series of patches to support using different
academic year and semester for the timetable builder.

We use a route format similar to NUSMods, like this:

    YYYY-YYYY/semX/

where `YYYY-YYYY` is the academic year, such as 2013-2014, and `semX` is
the semester, such as sem2.

Other changes required:
- For the `index.html` page, we prefix all the css and javascript paths
  with a '/'. This tells compojure to look at the `resources/public`
  folder for those files, instead of looking at the YYYY-YYYY/semX
  folder.
- Editing the urlPrefix for Exhibit3. For details, see the diff for this
  commit. The file is `src/js/vendor/exhibit3/exhibit-api.js`.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>